+++ src/components/InvestmentAccountList/InvestmentAccountList.tsx
@@ -1,13 +1,12 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import styled from 'styled-components';
-import { useQueryClient } from '@tanstack/react-query';
 
-import { Table, TableBody, TableContainer, Paper } from '@mui/material';
+import { Table, TableContainer, Paper } from '@mui/material';
 
 import InvestmentAccountItem from './InvestmentAccountItem/InvestmentAccountItem';
 import { useGetAccountQuery } from '@src/components/InvestmentAccountList/Account-query/InvestmentAccount.query';
 import { useGetFilteredAccountQuery } from './Account-query/FilteredInvestmentAccount.query';
-import InvestmentAccountRepository from './Account-query/InvestmentAccount.repository';
+import usePrefetchAccountList from './hooks/usePrefetchAccountList';
 import PagenationButton from './component/PagenationButton';
 import FilterButton from './component/FilterButton';
 import InvestmentAccountTableHead from './InvestmentAccountTableHead/InvestmentAccountTableHead';
@@ -17,7 +16,7 @@ const maxPage = 18;
 const InvestmentAccountList = () => {
   const [currentPage, setCurrentPage] = useState(1);
   const [keyword, setKeyword] = useState('');
-  
+
   const {
     data: defaultAccountListData,
     isLoading,
@@ -26,19 +25,7 @@ const InvestmentAccountList = () => {
 
   const { data: filteredAccountListData } = useGetFilteredAccountQuery(keyword);
 
-  const queryClient = useQueryClient();
-
-  useEffect(() => {
-    if (currentPage < maxPage) {
-      const nextPage = currentPage + 1;
-      queryClient.prefetchQuery(['GetInvestmentAccount', nextPage], () => {
-        return InvestmentAccountRepository.getInvestmentAccount(
-          nextPage,
-          maxPage
-        );
-      });
-    }
-  }, [currentPage, queryClient]);
+  usePrefetchAccountList(currentPage, maxPage);
 
   const handleCurrentPagePlus = () => {
     setCurrentPage(prev => prev + 1);
